+++ R/breaks-by-width.R
@@ -9,6 +9,8 @@
 #'
 #' @param start A scalar of class [Date][base::Dates] or [POSIXct][DateTimeClasses].
 #'   Can be omitted.
+#' @param shrink_last Logical. If `TRUE`, shrink the last interval to the last
+#'   finite value in `x`.
 #'
 #' @details
 #' If `width` is a Period, [`lubridate::add_with_rollback()`][`lubridate::m+`]
@@ -29,12 +31,12 @@ NULL
 #' @rdname chop_width
 #' @export
 #' @order 2
-brk_width <- function (width, start) UseMethod("brk_width")
+brk_width <- function (width, start, shrink_last = FALSE) UseMethod("brk_width")
 
 
 #' @rdname brk_width-for-datetime
 #' @export
-brk_width.Duration <- function (width, start) {
+brk_width.Duration <- function (width, start, shrink_last = FALSE) {
   loadNamespace("lubridate")
   width <- lubridate::make_difftime(as.numeric(width))
   NextMethod()
@@ -44,8 +46,9 @@ brk_width.Duration <- function (width, start) {
 #' @rdname chop_width
 #' @export
 #' @order 2
-brk_width.default <- function (width, start) {
+brk_width.default <- function (width, start, shrink_last = FALSE) {
   assert_that(is.scalar(width))
+  assert_that(is.flag(shrink_last), ! is.na(shrink_last))
 
   sm <- missing(start)
   if (! sm) assert_that(is.scalar(start))
@@ -66,6 +69,7 @@ brk_width.default <- function (width, start) {
       return(empty_breaks())
     }
 
+    if (shrink_last) breaks[length(breaks)] <- until
     if (sign(width) <= 0) breaks <- rev(breaks)
 
     breaks <- create_extended_breaks(breaks, x, extend, left, close_end)