+++ R/chop-by-width.R
@@ -6,6 +6,8 @@
 #' @param width Width of intervals.
 #' @param start Starting point for intervals. By default the smallest
 #'   finite `x` (largest if `width` is negative).
+#' @param shrink_last Logical. If `TRUE`, shrink the last interval to the last
+#'   finite value in `x`.
 #' @inheritParams chop
 #' @inherit chop-doc params return
 #'
@@ -24,6 +26,8 @@
 #'
 #' chop_width(1:10, 2, start = 0)
 #'
+#' chop_width(1:8, 2, shrink_last = TRUE)
+#'
 #' chop_width(1:9, -2)
 #'
 #' chop(1:10, brk_width(2, 0))
@@ -33,9 +37,10 @@ chop_width <- function (
                 width,
                 start,
                 ...,
-                left = sign(width) > 0
+                left = sign(width) > 0,
+                shrink_last = FALSE
               ) {
-  chop(x, brk_width(width, start), ..., left = left)
+  chop(x, brk_width(width, start, shrink_last = shrink_last), ..., left = left)
 }